+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `t-reduced` as a better wrapper than `make-transducers-reduced`.
 - The `t-median` reducer.
+- The `t-reversed` source to iterate from the end of an array/vector/string.
 
 #### Changed